+++ relay-metrics/src/cogs.rs
@@ -1,9 +1,8 @@
 //! COGS related metric utilities.
 
-use relay_base_schema::metrics::{MetricNamespace, MetricResourceIdentifier};
 use relay_cogs::{AppFeature, FeatureWeights};
 
-use crate::{Bucket, BucketView};
+use crate::{Bucket, BucketView, MetricNamespace};
 
 /// COGS estimator based on the estimated size of each bucket in bytes.
 pub struct BySize<'a>(pub &'a [Bucket]);
@@ -15,19 +14,22 @@ impl<'a> From<BySize<'a>> for FeatureWeights {
 }
 
 /// COGS estimator based on the bucket count.
-pub struct ByCount<'a>(pub &'a [Bucket]);
+pub struct ByCount<'a, T: IntoIterator<Item = &'a Bucket>>(pub T);
 
-impl<'a> From<ByCount<'a>> for FeatureWeights {
-    fn from(value: ByCount<'a>) -> Self {
+impl<'a, T: IntoIterator<Item = &'a Bucket>> From<ByCount<'a, T>> for FeatureWeights {
+    fn from(value: ByCount<'a, T>) -> Self {
         metric_app_features(value.0, |_| 1)
     }
 }
 
-fn metric_app_features(buckets: &[Bucket], f: impl Fn(&Bucket) -> usize) -> FeatureWeights {
+fn metric_app_features<'a, T>(buckets: T, f: impl Fn(&Bucket) -> usize) -> FeatureWeights
+where
+    T: IntoIterator<Item = &'a Bucket>,
+{
     let mut b = FeatureWeights::builder();
 
-    for bucket in buckets {
-        b.add_weight(to_app_feature(namespace(bucket)), f(bucket));
+    for bucket in buckets.into_iter() {
+        b.add_weight(to_app_feature(bucket.name.namespace()), f(bucket));
     }
 
     b.build()
@@ -43,10 +45,4 @@ fn to_app_feature(ns: MetricNamespace) -> AppFeature {
         MetricNamespace::Stats => AppFeature::MetricsStats,
         MetricNamespace::Unsupported => AppFeature::MetricsUnsupported,
     }
-}
-
-fn namespace(bucket: &Bucket) -> MetricNamespace {
-    MetricResourceIdentifier::parse(&bucket.name)
-        .map(|mri| mri.namespace)
-        .unwrap_or(MetricNamespace::Unsupported)
 }