+++ relay-server/src/metrics_extraction/transactions/mod.rs
@@ -5,7 +5,8 @@ use relay_common::time::UnixTimestamp;
 use relay_dynamic_config::{TagMapping, TransactionMetricsConfig};
 use relay_event_normalization::utils as normalize_utils;
 use relay_event_schema::protocol::{
-    AsPair, BrowserContext, Event, OsContext, TraceContext, TransactionSource,
+    AsPair, BrowserContext, Event, OsContext, PerformanceScoreContext, TraceContext,
+    TransactionSource,
 };
 use relay_metrics::{Bucket, DurationUnit, FiniteF64};
 
@@ -322,6 +323,9 @@ impl TransactionExtractor<'_> {
                     } else {
                         tags.clone()
                     },
+                    score_profile_version: is_performance_score
+                        .then(|| event.context::<PerformanceScoreContext>())
+                        .and_then(|context| context?.score_profile_version.value().cloned()),
                 };
 
                 metrics.project_metrics.push(
@@ -599,6 +603,7 @@ mod tests {
                             optional: false,
                         }],
                         condition: Some(RuleCondition::all()),
+                        version: Some("alpha".into()),
                     }],
                 }),
                 ..Default::default()
@@ -665,7 +670,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.foo@none",
+                name: MetricName(
+                    "d:transactions/measurements.foo@none",
+                ),
                 value: Distribution(
                     [
                         420.69,
@@ -692,7 +699,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.lcp@millisecond",
+                name: MetricName(
+                    "d:transactions/measurements.lcp@millisecond",
+                ),
                 value: Distribution(
                     [
                         3000.0,
@@ -720,7 +729,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.score.lcp@ratio",
+                name: MetricName(
+                    "d:transactions/measurements.score.lcp@ratio",
+                ),
                 value: Distribution(
                     [
                         0.0,
@@ -731,6 +742,7 @@ mod tests {
                     "environment": "fake_environment",
                     "geo.country_code": "US",
                     "release": "1.2.3",
+                    "sentry.score_profile_version": "alpha",
                     "transaction": "gEt /api/:version/users/",
                     "transaction.op": "mYOp",
                 },
@@ -741,7 +753,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.score.total@ratio",
+                name: MetricName(
+                    "d:transactions/measurements.score.total@ratio",
+                ),
                 value: Distribution(
                     [
                         0.0,
@@ -752,6 +766,7 @@ mod tests {
                     "environment": "fake_environment",
                     "geo.country_code": "US",
                     "release": "1.2.3",
+                    "sentry.score_profile_version": "alpha",
                     "transaction": "gEt /api/:version/users/",
                     "transaction.op": "mYOp",
                 },
@@ -762,7 +777,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.score.weight.lcp@ratio",
+                name: MetricName(
+                    "d:transactions/measurements.score.weight.lcp@ratio",
+                ),
                 value: Distribution(
                     [
                         1.0,
@@ -773,6 +790,7 @@ mod tests {
                     "environment": "fake_environment",
                     "geo.country_code": "US",
                     "release": "1.2.3",
+                    "sentry.score_profile_version": "alpha",
                     "transaction": "gEt /api/:version/users/",
                     "transaction.op": "mYOp",
                 },
@@ -783,7 +801,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/breakdowns.span_ops.ops.react.mount@millisecond",
+                name: MetricName(
+                    "d:transactions/breakdowns.span_ops.ops.react.mount@millisecond",
+                ),
                 value: Distribution(
                     [
                         2000.0,
@@ -810,7 +830,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "c:transactions/usage@none",
+                name: MetricName(
+                    "c:transactions/usage@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -822,7 +844,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration@millisecond",
+                name: MetricName(
+                    "d:transactions/duration@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -849,7 +873,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration_light@millisecond",
+                name: MetricName(
+                    "d:transactions/duration_light@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -866,7 +892,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "s:transactions/user@none",
+                name: MetricName(
+                    "s:transactions/user@none",
+                ),
                 value: Set(
                     {
                         933084975,
@@ -934,7 +962,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.fcp@millisecond",
+                name: MetricName(
+                    "d:transactions/measurements.fcp@millisecond",
+                ),
                 value: Distribution(
                     [
                         1.1,
@@ -953,7 +983,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.foo@none",
+                name: MetricName(
+                    "d:transactions/measurements.foo@none",
+                ),
                 value: Distribution(
                     [
                         8.8,
@@ -971,7 +1003,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.stall_count@none",
+                name: MetricName(
+                    "d:transactions/measurements.stall_count@none",
+                ),
                 value: Distribution(
                     [
                         3.3,
@@ -989,7 +1023,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "c:transactions/usage@none",
+                name: MetricName(
+                    "c:transactions/usage@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -1001,7 +1037,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration@millisecond",
+                name: MetricName(
+                    "d:transactions/duration@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -1019,7 +1057,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration_light@millisecond",
+                name: MetricName(
+                    "d:transactions/duration_light@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -1073,7 +1113,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.fcp@second",
+                name: MetricName(
+                    "d:transactions/measurements.fcp@second",
+                ),
                 value: Distribution(
                     [
                         1.1,
@@ -1092,7 +1134,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/measurements.lcp@none",
+                name: MetricName(
+                    "d:transactions/measurements.lcp@none",
+                ),
                 value: Distribution(
                     [
                         2.2,
@@ -1111,7 +1155,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "c:transactions/usage@none",
+                name: MetricName(
+                    "c:transactions/usage@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -1123,7 +1169,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration@millisecond",
+                name: MetricName(
+                    "d:transactions/duration@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -1141,7 +1189,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration_light@millisecond",
+                name: MetricName(
+                    "d:transactions/duration_light@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -1269,7 +1319,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/measurements.a_custom1@none",
+                name: MetricName(
+                    "d:transactions/measurements.a_custom1@none",
+                ),
                 value: Distribution(
                     [
                         41.0,
@@ -1287,7 +1339,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/measurements.fcp@millisecond",
+                name: MetricName(
+                    "d:transactions/measurements.fcp@millisecond",
+                ),
                 value: Distribution(
                     [
                         0.123,
@@ -1306,7 +1360,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/measurements.g_custom2@second",
+                name: MetricName(
+                    "d:transactions/measurements.g_custom2@second",
+                ),
                 value: Distribution(
                     [
                         42.0,
@@ -1324,7 +1380,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "c:transactions/usage@none",
+                name: MetricName(
+                    "c:transactions/usage@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -1336,7 +1394,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/duration@millisecond",
+                name: MetricName(
+                    "d:transactions/duration@millisecond",
+                ),
                 value: Distribution(
                     [
                         2000.0,
@@ -1354,7 +1414,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/duration_light@millisecond",
+                name: MetricName(
+                    "d:transactions/duration_light@millisecond",
+                ),
                 value: Distribution(
                     [
                         2000.0,
@@ -1505,7 +1567,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "c:transactions/usage@none",
+                name: MetricName(
+                    "c:transactions/usage@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -1517,7 +1581,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration@millisecond",
+                name: MetricName(
+                    "d:transactions/duration@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -1535,7 +1601,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "d:transactions/duration_light@millisecond",
+                name: MetricName(
+                    "d:transactions/duration_light@millisecond",
+                ),
                 value: Distribution(
                     [
                         59000.0,
@@ -1663,7 +1731,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420400),
                 width: 0,
-                name: "c:transactions/count_per_root_project@none",
+                name: MetricName(
+                    "c:transactions/count_per_root_project@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -1930,16 +2000,36 @@ mod tests {
 
         insta::assert_debug_snapshot!(metrics_names, @r###"
         [
-            "d:transactions/measurements.frames_frozen@none",
-            "d:transactions/measurements.frames_frozen_rate@ratio",
-            "d:transactions/measurements.frames_slow@none",
-            "d:transactions/measurements.frames_slow_rate@ratio",
-            "d:transactions/measurements.frames_total@none",
-            "d:transactions/measurements.stall_percentage@ratio",
-            "d:transactions/measurements.stall_total_time@millisecond",
-            "c:transactions/usage@none",
-            "d:transactions/duration@millisecond",
-            "d:transactions/duration_light@millisecond",
+            MetricName(
+                "d:transactions/measurements.frames_frozen@none",
+            ),
+            MetricName(
+                "d:transactions/measurements.frames_frozen_rate@ratio",
+            ),
+            MetricName(
+                "d:transactions/measurements.frames_slow@none",
+            ),
+            MetricName(
+                "d:transactions/measurements.frames_slow_rate@ratio",
+            ),
+            MetricName(
+                "d:transactions/measurements.frames_total@none",
+            ),
+            MetricName(
+                "d:transactions/measurements.stall_percentage@ratio",
+            ),
+            MetricName(
+                "d:transactions/measurements.stall_total_time@millisecond",
+            ),
+            MetricName(
+                "c:transactions/usage@none",
+            ),
+            MetricName(
+                "d:transactions/duration@millisecond",
+            ),
+            MetricName(
+                "d:transactions/duration_light@millisecond",
+            ),
         ]
         "###);
     }
@@ -2001,7 +2091,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/measurements.lcp@millisecond",
+                name: MetricName(
+                    "d:transactions/measurements.lcp@millisecond",
+                ),
                 value: Distribution(
                     [
                         41.0,
@@ -2018,7 +2110,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "c:transactions/usage@none",
+                name: MetricName(
+                    "c:transactions/usage@none",
+                ),
                 value: Counter(
                     1.0,
                 ),
@@ -2030,7 +2124,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/duration@millisecond",
+                name: MetricName(
+                    "d:transactions/duration@millisecond",
+                ),
                 value: Distribution(
                     [
                         2000.0,
@@ -2047,7 +2143,9 @@ mod tests {
             Bucket {
                 timestamp: UnixTimestamp(1619420402),
                 width: 0,
-                name: "d:transactions/duration_light@millisecond",
+                name: MetricName(
+                    "d:transactions/duration_light@millisecond",
+                ),
                 value: Distribution(
                     [
                         2000.0,