+++ relay-server/src/utils/bucket_encoding.rs
@@ -1,9 +1,7 @@
 use std::io;
 
 use relay_dynamic_config::{BucketEncoding, GlobalConfig};
-use relay_metrics::{
-    Bucket, BucketValue, FiniteF64, MetricNamespace, MetricResourceIdentifier, SetView,
-};
+use relay_metrics::{Bucket, BucketValue, FiniteF64, MetricNamespace, SetView};
 use serde::Serialize;
 
 static BASE64: data_encoding::Encoding = data_encoding::BASE64;
@@ -30,9 +28,7 @@ impl<'a> BucketEncoder<'a> {
     /// afterwards the bucket can be split into multiple smaller views
     /// and encoded one by one.
     pub fn prepare(&self, bucket: &mut Bucket) -> MetricNamespace {
-        let namespace = MetricResourceIdentifier::parse(&bucket.name)
-            .map(|mri| mri.namespace)
-            .unwrap_or(MetricNamespace::Unsupported);
+        let namespace = bucket.name.namespace();
 
         if let BucketValue::Distribution(ref mut distribution) = bucket.value {
             let enc = self.global_config.options.metric_bucket_dist_encodings;